refactor: Removed outdated code structure
