+++ database/populators/agents.json
@@ -4,8 +4,8 @@
       "name": "Gateway Orchestrator",
       "description": "Central orchestrator that standardizes and routes external RPA error notifications.",
       "prompt": null,
-      "prompt_import": "gateway.prompts:GATEWAY_ORCHESTRATOR_PROMPT",
-      "response_model": "gateway.templates.ResponseToRPA",
+      "prompt_import": "prompts.gateway:GATEWAY_ORCHESTRATOR_PROMPT",
+      "response_model": "templates.gateway.ResponseToRPA",
       "router_model": "$PROVIDER_MODEL",
       "agent_type": "GatewayAgent",
       "input_type": "TEXT",
@@ -39,8 +39,8 @@
       "name": "UI Exception Handler",
       "description": "Module responsible for managing UI-related errors. This module has access to the current UI state via screenshots and is provided with visual information about the moment where execution failed. It specializes in identifying and resolving issues related to user interface interactions, element detection, and visual validation failures.",
       "prompt": null,
-      "prompt_import": "modules.uierror.prompts:UI_EXCEPTION_HANDLER",
-      "response_model": "modules.uierror.templates.UiExceptionReport",
+      "prompt_import": "prompts.error_handler:UI_EXCEPTION_HANDLER",
+      "response_model": "templates.error_handler.UiExceptionReport",
       "router_model": "$PROVIDER_MODEL",
       "agent_type": "ErrorAgent",
       "input_type": "TEXT",
@@ -86,8 +86,8 @@
       "name": "UI Direct Recovery",
       "description": "Performs immediate direct UI recovery using vision and tool models.",
       "prompt": null,
-      "prompt_import": "modules.uierror.prompts:RECOVERY_DIRECT_PROMPT",
-      "response_model": "modules.uierror.templates.RecoveryDirectReport",
+      "prompt_import": "prompts.gui_agent:RECOVERY_DIRECT_PROMPT",
+      "response_model": "templates.gui_agent.RecoveryDirectReport",
       "router_model": "$PROVIDER_VISION_TOOL_MODEL",
       "agent_type": "GuiAgent",
       "input_type": "IMAGETEXT",
@@ -127,8 +127,8 @@
       "name": "UI Recovery Planner",
       "description": "Generates a multi-step recovery plan for a UI failure.",
       "prompt": null,
-      "prompt_import": "modules.uierror.prompts:RECOVERY_PLANNER_PROMPT",
-      "response_model": "modules.uierror.templates.RecoveryPlannerReport",
+      "prompt_import": "prompts.gui_agent:RECOVERY_PLANNER_PROMPT",
+      "response_model": "templates.gui_agent.RecoveryPlannerReport",
       "router_model": "$PROVIDER_VISION_MODEL",
       "agent_type": "GuiAgent",
       "input_type": "IMAGETEXT",
@@ -174,8 +174,8 @@
       "name": "UI Step Executor",
       "description": "Executes individual planned recovery steps and validates outcomes.",
       "prompt": null,
-      "prompt_import": "modules.uierror.prompts:RECOVERY_STEP_EXECUTION_PROMPT",
-      "response_model": "modules.uierror.templates.RecoveryStepExecutionResult",
+      "prompt_import": "prompts.gui_agent:RECOVERY_STEP_EXECUTION_PROMPT",
+      "response_model": "templates.gui_agent.RecoveryStepExecutionResult",
       "router_model": "$PROVIDER_VISION_TOOL_MODEL",
       "agent_type": "GuiAgent",
       "input_type": "IMAGETEXT",
