+++ prompts/__init__.py
@@ -1,0 +1,21 @@
+from prompts.error_handler import UI_EXCEPTION_HANDLER
+from prompts.gateway import GATEWAY_ORCHESTRATOR_PROMPT
+from prompts.gui_agent import (
+    COMPUTER_USE_DOUBAO,
+    RECOVERY_ACTION_PROMPT,
+    RECOVERY_DIRECT_PROMPT,
+    RECOVERY_PLANNER_PROMPT,
+    RECOVERY_STEP_EXECUTION_PROMPT,
+    STANDALONE_COMPUTER_USE_DOUBAO,
+)
+
+__all__ = [
+    "COMPUTER_USE_DOUBAO",
+    "GATEWAY_ORCHESTRATOR_PROMPT",
+    "RECOVERY_ACTION_PROMPT",
+    "RECOVERY_DIRECT_PROMPT",
+    "RECOVERY_PLANNER_PROMPT",
+    "RECOVERY_STEP_EXECUTION_PROMPT",
+    "STANDALONE_COMPUTER_USE_DOUBAO",
+    "UI_EXCEPTION_HANDLER",
+]
