+++ prompts/gateway.py
@@ -1,5 +1,3 @@
-# Gateway Agent Prompts for RPA Recovery Framework
-
 GATEWAY_ORCHESTRATOR_PROMPT = """
 You are the Gateway Agent, the central orchestrator for the RPA Recovery Framework. Your primary responsibility is to receive error notifications from external RPA systems, analyze them, and intelligently route them to the most appropriate recovery module.
