+++ prompts/gui_agent.py
@@ -1,8 +1,3 @@
-from settings import UI_ERROR_PLANNING, UI_MID_AGENT
-
-# and whether you expect a ui change to be visible. Note that some action may trigger it and some may not, this is to ensure the action was correctly executed. Things like writting text on inputs or checking boxes will most likely not trigger a noticabl UI change.
-
-# Custom system prompts for the RPA recovery scenario
 RECOVERY_DIRECT_PROMPT = """
 You are a specialized AI agent designed to recover robotic process automation (RPA) workflows that have failed.
 Your role is to analyze the current state, understand what went wrong, and iteratively interact with the screen to get the process back on track.
@@ -109,54 +104,6 @@ Example steps could include:
 - Fill in form fields
 """
 
-UI_EXCEPTION_HANDLER = f"""
-You are a specialized AI agent designed to recover robotic process automation (RPA) workflows that have failed.
-Your role is to analyze the current state, understand what went wrong, create, and execute a plan to get the process back on track.
-
-You will be given:
-1. The previous successful actions performed by the robot
-2. The action that was expected to be performed but failed (failedActivity, pay special attention to this)
-3. Information about the overall process
-4. A list of variables used in the process, including the ones that may have already been used. If you need to use them, include their values in the plan.
-
-Follow these guidelines:
-{
-    '''
-1. Use tools at your disposal to generate a recovery plan, do not generate it yourself
-2. As the task name, provide a short description of the final task (e.g., "Login to the application", "Obtain weather data", etc.)
-3. After a plan is generated, execute it step by step using the `step_execution_handler` tool.
-      '''
-    if UI_ERROR_PLANNING
-    else '''
-1. Use tools at your disposal to delegate the recovery actions, do not generate them yourself
-2. As the task name, provide a short description of the final task (e.g., "Login to the application", "Obtain weather data", etc.)
-3. Use the `recovery_agent` tool.
-      '''
-    if not UI_MID_AGENT
-    else '''
-1. Use tools at your disposal to delegate the recovery actions, do not generate them yourself
-2. As the task name, provide a short description of the final task (e.g., "Login to the application", "Obtain weather data", etc.)
-3. Use the `standalone_uitars` tool.
-      '''
-}
-
-After the recovery is executed, if it is succesful, identify which of the futureActivities have already been completed by the recovery process.
-You will do so by using the compute_continuation_activity tool, providing the list of futureActivities and a list of booleans with the same length indicating which futureActivity were executed during the recovery process.
-If the last futureActivity was executed, -1 is returned by the compute_continuation_activity to indicate the robot can finish its execution.
-
-Your final report, after executing all steps, should include the following:
-- Reasoning and Steps
-  - Failure analysis: "Analysis of what may have caused the failure"
-  - UI state: "Description of the current UI state and how it differs from expected"
-  - Recovery approach: "General approach for recovery"
-  - Challenges: "Potential challenges or alternative approaches"
-- Steps: ["Step 1", "Step 2", "Step 3", "..."]
-- Future Activities: {{"futureActivity 1": "executed on step X", "futureActivity 2": "not executed", "..."}}
-- Result: "The recovery plan was successfully executed."
-- Finished activity: True|False (If continue activity is -1, then True, else False)
-- Continue from step: <step number from where the robot should continue execution, using the compute_continuation_activity tool>
-"""
-
 RECOVERY_STEP_EXECUTION_PROMPT = """
 You are an AI agent designed to execute steps in a recovery plan for an rpa robot. Your purpose is to orchestrate the necessary actions to resolve a UI error and get the robot back on track.
 
@@ -198,10 +145,10 @@ left_double(point='<point>x1 y1</point>')
 right_single(point='<point>x1 y1</point>')
 drag(start_point='<point>x1 y1</point>', end_point='<point>x2 y2</point>')
 hotkey(key='ctrl c') # Split keys with a space and use lowercase. Also, do not use more than 3 keys in one hotkey action.
-type(content='xxx') # Use escape characters \\', \\\", and \\n in content part to ensure we can parse the content in normal python string format. If you want to submit your input, use \\n at the end of content.
+type(content='xxx') # Use escape characters \\', \\\" and \\n in content part to ensure we can parse the content in normal python string format. If you want to submit your input, use \\n at the end of content.
 scroll(point='<point>x1 y1</point>', direction='down or up or right or left') # Show more information on the `direction` side.
 wait() #Sleep for 5s and take a screenshot to check for any changes.
-finished(content='xxx') # Use escape characters \\', \\", and \\n in content part to ensure we can parse the content in normal python string format.
+finished(content='xxx') # Use escape characters \\', \\" and \\n in content part to ensure we can parse the content in normal python string format.
 
 
 ## Note
@@ -247,10 +194,10 @@ left_double(point='<point>x1 y1</point>')
 right_single(point='<point>x1 y1</point>')
 drag(start_point='<point>x1 y1</point>', end_point='<point>x2 y2</point>')
 hotkey(key='ctrl c') # Split keys with a space and use lowercase. Also, do not use more than 3 keys in one hotkey action.
-type(content='xxx') # Use escape characters \\', \\\", and \\n in content part to ensure we can parse the content in normal python string format. If you want to submit your input, use \\n at the end of content.
+type(content='xxx') # Use escape characters \\', \\\" and \\n in content part to ensure we can parse the content in normal python string format. If you want to submit your input, use \\n at the end of content.
 scroll(point='<point>x1 y1</point>', direction='down or up or right or left') # Show more information on the `direction` side.
 wait() #Sleep for 5s and take a screenshot to check for any changes.
-finished(content='xxx') # Use escape characters \\', \\", and \\n in content part to ensure we can parse the content in normal python string format.
+finished(content='xxx') # Use escape characters \\', \\" and \\n in content part to ensure we can parse the content in normal python string format.
 
 
 ## Note
