+++ templates/__init__.py
@@ -1,0 +1,27 @@
+from templates.error_handler import UiExceptionReport
+from templates.gateway import ResponseToRPA
+from templates.gui_agent import (
+    RecoveryActionDetail,
+    RecoveryActionPayload,
+    RecoveryDirectReport,
+    RecoveryPlannerReport,
+    RecoveryReasoning,
+    RecoveryStepExecutionResult,
+    TemplateDefinition,
+    TemplateModel,
+    TEMPLATES,
+)
+
+__all__ = [
+    "RecoveryActionDetail",
+    "RecoveryActionPayload",
+    "RecoveryDirectReport",
+    "RecoveryPlannerReport",
+    "RecoveryReasoning",
+    "RecoveryStepExecutionResult",
+    "ResponseToRPA",
+    "TemplateDefinition",
+    "TemplateModel",
+    "TEMPLATES",
+    "UiExceptionReport",
+]
