+++ templates/common.py
@@ -1,0 +1,17 @@
+from __future__ import annotations
+
+import json
+from typing import override
+
+from pydantic import BaseModel
+
+
+class TemplateModel(BaseModel):
+    """Base model that provides consistent formatting helpers."""
+
+    class ConfigDict:
+        extra = "forbid"
+
+    @override
+    def __str__(self) -> str:  # pragma: no cover - simple serialization helper
+        return json.dumps(self.model_dump(), indent=2, ensure_ascii=True)
